#18 MOVE handle mouse clicks in CrypterScene
